The return
took a break from coding after bootcamp. Starting back at the basics this time adding some books in the mix.
This tale will be sporadic but I will try to comment and note transitions from book to book. Current book is Decoding JavaScript : A simple Guide for the not-so-simple yada yada yada...
This is purely for reflection. I may delete this repo at any point =)